perf: skip aliases
When this plugin is used in combination with nodenv-aliases, some
aliases are created that match the regexp `^[0-9.]+$` and are therefore
passed on to sh-semver which seems particularily slow at dealing with
them.

Since aliases are redundant anyway, we can simply ask nodenv to skip
them when listing installed node versions.

Fixes: Significant slowdown when this plugin is active; node starts 2-seconds slower #8